Rename feature/module "log" to "logs".
