+++ opentelemetry-api/Cargo.toml
@@ -29,4 +29,4 @@ default = ["trace"]
 trace = ["pin-project-lite"]
 metrics = ["fnv"]
 testing = ["trace"]
-log = []
+logs = []
