+++ opentelemetry/Cargo.toml
@@ -28,7 +28,7 @@ opentelemetry_sdk = { version = "0.19", path = "../opentelemetry-sdk" }
 default = ["trace"]
 trace = ["opentelemetry_api/trace", "opentelemetry_sdk/trace"]
 metrics = ["opentelemetry_api/metrics", "opentelemetry_sdk/metrics"]
-log = ["opentelemetry_sdk/log"]
+logs = ["opentelemetry_sdk/logs"]
 testing = ["opentelemetry_api/testing", "opentelemetry_sdk/testing"]
 rt-tokio = ["opentelemetry_sdk/rt-tokio"]
 rt-tokio-current-thread = ["opentelemetry_sdk/rt-tokio-current-thread"]
